+++ src/TaskEither.ts
@@ -463,15 +463,6 @@ declare module './HKT' {
 }
 
 /**
- * Semigroup returning the left-most non-`Left` value. If both operands are `Right`s then the inner values are
- * concatenated using the provided `Semigroup`
- *
- * @category instances
- * @since 3.0.0
- */
-export const getSemigroup = <A, E>(S: Semigroup<A>): Semigroup<TaskEither<E, A>> => T.getSemigroup(E.getSemigroup(S))
-
-/**
  * @category instances
  * @since 3.0.0
  */